Implement untaring in rust
This avoids relying on the system path

The exposed `untar` attribute is pretty ugly, but I'm not sure the best
alternative. We need to be able to access `untar` from anything which
has an `out_dir_tar` attribute, which may be any of our rules, but need
to specifically not depend on it from the targets which make `untar`
itself. This means it can't live in a toolchain.

Ideally it'd be an attribute conditionally set only iff `out_dir_tar` is
set, which we could do in a macro, but introducing a layer of macros to
the rules just to make our handful of bootstrap rust_libraries not need
to fake out the attribute seems overkill. The binary will only
_actually_ be depended on and built if it's needed.